Add a legacy excerpts output for events
I'm calling this events_legacy so that we're free to create a better
version with some more thought in the future for the new web
framework. This will work as a stop-gap until then and can be moved
to a legacy plugin for sites that use it in the future.